Enum version_compare::version_part::VersionPart

pub enum VersionPart<'a> {
    Number(i32),
    Text(&'a str),
}

Enum of version string parts.

Each version string is broken down into these version parts when being parsed to a Version.

Variants
Number(i32)

Numeric part, most common in version strings. Holds the numerical value.

Text(&'a str)

A text part. These parts usually hold text with an yet unknown definition. Holds the string slice.